@@ -11,14 +11,14 @@ using MailKit.Security;
 
 namespace Check_NOAA_Outlook
 {
-    public class MainProgram
+    public partial class MainProgram
     {
         public MainProgram()
         {
             m_strCD = Environment.CurrentDirectory;
-            if (!m_strCD.EndsWith("\\")) m_strCD += "\\";
+            if (!m_strCD.EndsWith('\\')) m_strCD += "\\";
 
-            m_Locations = new List<LocationsInfo>();
+            m_Locations = [];
 
             m_SourceCodes = new string[3];
             //m_GetExtendedOutlook = true;
@@ -50,40 +50,43 @@ namespace Check_NOAA_Outlook
         private bool m_EmailIsEnabled, m_IncludeMeso; // m_GetExtendedOutlook
         private System.Net.Http.HttpClient m_MainClient;
         private int m_LocationsInSevere, m_LocationsEmailed, m_MesosIncluded;
+        private static readonly Regex RegexTo = new(@"\(.*\)");
+        private static readonly Regex RegexTitle = new(".*title=.*\"Categorical Outlook\"");
+        private static readonly Regex RegexOnClick = new(@"\((.*?)\)");
 
         private bool LoadSettings()
         {
-            string AppSettingsFile = System.IO.Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "appsettings.json");
+            string appSettingsFile = System.IO.Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "appsettings.json");
 
-            if (!System.IO.File.Exists(AppSettingsFile))
+            if (!System.IO.File.Exists(appSettingsFile))
             {
-                WriteToConsoleAndLogIt(WhichLogType.FATAL, "File not found: " + AppSettingsFile);
+                WriteToConsoleAndLogIt(WhichLogType.FATAL, "File not found: " + appSettingsFile);
                 return false;
             }
 
-            string AppSettingsText;
+            string appSettingsText;
             try
-            { AppSettingsText = System.IO.File.ReadAllText(AppSettingsFile); }
+            { appSettingsText = System.IO.File.ReadAllText(appSettingsFile); }
             catch (Exception ex)
             {
-                WriteToConsoleAndLogItWithException(WhichLogType.FATAL, "Unable to read " + AppSettingsFile, ex);
+                WriteToConsoleAndLogItWithException(WhichLogType.FATAL, "Unable to read " + appSettingsFile, ex);
                 return false;
             }
 
-            Dictionary<string, object> AppSettings;
+            Dictionary<string, object> appSettings;
             try
             {
-                AppSettings = new Dictionary<string, object>(
+                appSettings = new Dictionary<string, object>(
                     System.Text.Json.JsonSerializer.Deserialize<Dictionary<string, object>>(
-                    AppSettingsText, m_JsonOpts), StringComparer.CurrentCultureIgnoreCase);
+                    appSettingsText, m_JsonOpts), StringComparer.CurrentCultureIgnoreCase);
             }
             catch (Exception ex)
             {
-                WriteToConsoleAndLogItWithException(WhichLogType.FATAL, "Unable to deserialize config file " + AppSettingsFile, ex);
+                WriteToConsoleAndLogItWithException(WhichLogType.FATAL, "Unable to deserialize config file " + appSettingsFile, ex);
                 return false;
             }
 
-            string temp = (string)GetAppSetting("LoggingLevel", "INFO", AppSettings);
+            string temp = (string)GetAppSetting("LoggingLevel", "INFO", appSettings);
             if (string.IsNullOrEmpty(temp)) temp = "INFO";
             log.m_LogLevel = (temp.ToLower()) switch
             {
@@ -94,50 +97,50 @@ namespace Check_NOAA_Outlook
                 "warn" => WhichLogType.WARN,
                 _ => WhichLogType.INFO,
             };
-            string Temp = (string)GetAppSetting("LogFile", null, AppSettings);
-            if (!string.IsNullOrEmpty(Temp)) this.log.LogFile = Environment.ExpandEnvironmentVariables(Temp);
+            temp = (string)GetAppSetting("LogFile", null, appSettings);
+            if (!string.IsNullOrEmpty(temp)) this.log.LogFile = Environment.ExpandEnvironmentVariables(temp);
 
             AppDomain.CurrentDomain.SetData("LogDir", System.IO.Path.GetDirectoryName(this.log.LogFile));
 
-            int MaxLogSizeMB = Convert.ToInt32(GetAppSetting("LogMaxFileSizeMB", 0, AppSettings));
-            int MaxLogRotations = Convert.ToInt32(GetAppSetting("LogFilesToKeep", 0, AppSettings));
+            int maxLogSizeMB = Convert.ToInt32(GetAppSetting("LogMaxFileSizeMB", 0, appSettings));
+            int maxLogRotations = Convert.ToInt32(GetAppSetting("LogFilesToKeep", 0, appSettings));
 
-            if (MaxLogSizeMB > 0)
+            if (maxLogSizeMB > 0)
             {
-                bool DidRotate = this.log.CheckIfRotationNeeded(MaxLogRotations, MaxLogSizeMB, out string RotationMsg);
-                if (DidRotate)
+                bool didRotate = this.log.CheckIfRotationNeeded(maxLogRotations, maxLogSizeMB, out string RotationMsg);
+                if (didRotate)
                     log.Info("Log rotation: " + RotationMsg);
                 else if (RotationMsg != null)
                     log.Debug("Log rotation: " + RotationMsg);
             }
 
-            string VersionStr = System.Reflection.Assembly.GetExecutingAssembly().GetName().Version.ToString();
-            log.InfoFormat("------------------------------ Program starting.  Version {0} -----------------------------------------", VersionStr);
+            string versionStr = System.Reflection.Assembly.GetExecutingAssembly().GetName().Version.ToString();
+            log.InfoFormat("------------------------------ Program starting.  Version {0} -----------------------------------------", versionStr);
 
-            m_EmailHost = (string)GetAppSetting("Email_Host", "localhost", AppSettings);
-            m_EmailUser = (string)GetAppSetting("Email_Username", null, AppSettings);
-            m_EmailPassword = (string)GetAppSetting("Email_Password", null, AppSettings);
-            m_EmailPort = Convert.ToInt32(GetAppSetting("Email_Port", 587, AppSettings));
-            m_UseSSL = Convert.ToBoolean(GetAppSetting("Email_UseSSL", false, AppSettings));
-            m_EmailFrom = (string)GetAppSetting("Email_From", null, AppSettings);
-            m_EmailReplyTo = (string)GetAppSetting("Email_ReplyTo", null, AppSettings);
-            m_Email_HiPri_MinSev = Convert.ToInt32(GetAppSetting("Email_HighPri_Min_Criteria", "2", AppSettings));
-            m_EmailIsEnabled = Convert.ToBoolean(GetAppSetting("Email_IsEnabled", "true", AppSettings));
-            m_Download_MaxRetries = Convert.ToInt32(GetAppSetting("Download_Max_Retries", "3", AppSettings));
-            m_Download_Timeout = Convert.ToInt32(GetAppSetting("Download_Timeout_Seconds", "10", AppSettings));
-            m_NOAA_URL = (string)GetAppSetting("NOAA_Main_URL", "http://www.spc.noaa.gov/products/outlook/", AppSettings);
-            m_MesoBaseURL = (string)GetAppSetting("NOAA_Meso_URL", "http://www.spc.noaa.gov/products/outlook/", AppSettings);
+            m_EmailHost = (string)GetAppSetting("Email_Host", "localhost", appSettings);
+            m_EmailUser = (string)GetAppSetting("Email_Username", null, appSettings);
+            m_EmailPassword = (string)GetAppSetting("Email_Password", null, appSettings);
+            m_EmailPort = Convert.ToInt32(GetAppSetting("Email_Port", 587, appSettings));
+            m_UseSSL = Convert.ToBoolean(GetAppSetting("Email_UseSSL", false, appSettings));
+            m_EmailFrom = (string)GetAppSetting("Email_From", null, appSettings);
+            m_EmailReplyTo = (string)GetAppSetting("Email_ReplyTo", null, appSettings);
+            m_Email_HiPri_MinSev = Convert.ToInt32(GetAppSetting("Email_HighPri_Min_Criteria", "2", appSettings));
+            m_EmailIsEnabled = Convert.ToBoolean(GetAppSetting("Email_IsEnabled", "true", appSettings));
+            m_Download_MaxRetries = Convert.ToInt32(GetAppSetting("Download_Max_Retries", "3", appSettings));
+            m_Download_Timeout = Convert.ToInt32(GetAppSetting("Download_Timeout_Seconds", "10", appSettings));
+            m_NOAA_URL = (string)GetAppSetting("NOAA_Main_URL", "http://www.spc.noaa.gov/products/outlook/", appSettings);
+            m_MesoBaseURL = (string)GetAppSetting("NOAA_Meso_URL", "http://www.spc.noaa.gov/products/outlook/", appSettings);
             //m_EmailTestAddress = (string)GetAppSetting("Email_Test_Address", null, AppSettings);
-            m_IncludeMeso = Convert.ToBoolean(GetAppSetting("IncludeMesoDiscussions", "false", AppSettings));
+            m_IncludeMeso = Convert.ToBoolean(GetAppSetting("IncludeMesoDiscussions", "false", appSettings));
 
-            if (!m_NOAA_URL.EndsWith("/")) m_NOAA_URL += "/";
+            if (!m_NOAA_URL.EndsWith('/')) m_NOAA_URL += "/";
 
-            m_URLs = new List<string>
-            {
+            m_URLs =
+            [
                 m_NOAA_URL + "day1otlk.html",
                 m_NOAA_URL + "day2otlk.html",
                 m_NOAA_URL + "day3otlk.html"
-            };
+            ];
             if (m_IncludeMeso) m_URLs.Add(m_MesoBaseURL);
 
             //m_URL_Extended = "http://www.spc.noaa.gov/products/exper/day4-8/";
@@ -171,101 +174,101 @@ namespace Check_NOAA_Outlook
 
         private bool LoadJsonFiles()
         {
-            string LocationsFile = m_strCD + "Locations.json";
-            string TrackingFile = m_strCD + "Tracking.json";
+            string locationsFile = m_strCD + "Locations.json";
+            string trackingFile = m_strCD + "Tracking.json";
 
-            if (!System.IO.File.Exists(LocationsFile))
+            if (!System.IO.File.Exists(locationsFile))
             {
-                WriteToConsoleAndLogIt(WhichLogType.DEBUG, "File not found: " + LocationsFile);
+                WriteToConsoleAndLogIt(WhichLogType.DEBUG, "File not found: " + locationsFile);
                 return false;
             }
 
-            string FileContents;
+            string fileContents;
             try
-            { FileContents = System.IO.File.ReadAllText(LocationsFile); }
+            { fileContents = System.IO.File.ReadAllText(locationsFile); }
             catch (Exception ex)
             {
-                WriteToConsoleAndLogItWithException(WhichLogType.FATAL, "Unable to read " + LocationsFile, ex);
+                WriteToConsoleAndLogItWithException(WhichLogType.FATAL, "Unable to read " + locationsFile, ex);
                 return false;
             }
 
-            Dictionary<string, LocationsInfo> LocationsDictionary;
+            Dictionary<string, LocationsInfo> locationsDictionary;
             try
-            { LocationsDictionary = System.Text.Json.JsonSerializer.Deserialize<Dictionary<string, LocationsInfo>>(FileContents, m_JsonOpts); }
+            { locationsDictionary = System.Text.Json.JsonSerializer.Deserialize<Dictionary<string, LocationsInfo>>(fileContents, m_JsonOpts); }
             catch (Exception ex)
             {
-                WriteToConsoleAndLogItWithException(WhichLogType.FATAL, "Unable to deserialize " + LocationsFile, ex);
+                WriteToConsoleAndLogItWithException(WhichLogType.FATAL, "Unable to deserialize " + locationsFile, ex);
                 return false;
             }
 
-            if (LocationsDictionary == null || LocationsDictionary.Count == 0)
+            if (locationsDictionary == null || locationsDictionary.Count == 0)
             {
                 WriteToConsoleAndLogIt(WhichLogType.FATAL, "No locations were defined");
                 return false;
             }
 
-            foreach (string LocationId in LocationsDictionary.Keys)
+            foreach (string locationId in locationsDictionary.Keys)
             {
-                LocationsInfo LI = LocationsDictionary[LocationId];
-                LI.LocationId = LocationId;
+                LocationsInfo li = locationsDictionary[locationId];
+                li.locationId = locationId;
 
                 //LI.LocationId = (int)Row["Id"];
-                string Loc = LI.Coordinates;
-                string[] Points = Loc.Split(',');
-                if (Points.Length != 2)
+                string loc = li.Coordinates;
+                string[] points = loc.Split(',');
+                if (points.Length != 2)
                 {
-                    WriteToConsoleAndLogIt(WhichLogType.FATAL, "Invalid coordinates: " + Loc);
+                    WriteToConsoleAndLogIt(WhichLogType.FATAL, "Invalid coordinates: " + loc);
                     return false;
                 }
 
-                LI.Location = new System.Drawing.Point(
-                    Convert.ToInt32(Points[0]),
-                    Convert.ToInt32(Points[1])
+                li.location = new System.Drawing.Point(
+                    Convert.ToInt32(points[0]),
+                    Convert.ToInt32(points[1])
                     );
 
-                if (!string.IsNullOrWhiteSpace(LI.MesoMap_Coordinates))
+                if (!string.IsNullOrWhiteSpace(li.MesoMap_Coordinates))
                 {
-                    Points = LI.MesoMap_Coordinates.Split(',');
-                    if (Points.Length == 2)
+                    points = li.MesoMap_Coordinates.Split(',');
+                    if (points.Length == 2)
                     {
-                        LI.MesoMapLocation = new System.Drawing.Point(
-                            Convert.ToInt32(Points[0]),
-                            Convert.ToInt32(Points[1])
+                        li.mesoMapLocation = new System.Drawing.Point(
+                            Convert.ToInt32(points[0]),
+                            Convert.ToInt32(points[1])
                             );
                     }
                 }
 
-                m_Locations.Add(LI);
+                m_Locations.Add(li);
             }
 
-            if (!System.IO.File.Exists(TrackingFile))
+            if (!System.IO.File.Exists(trackingFile))
                 return true;
 
             try
-            { FileContents = System.IO.File.ReadAllText(TrackingFile); }
+            { fileContents = System.IO.File.ReadAllText(trackingFile); }
             catch (Exception ex)
             {
-                WriteToConsoleAndLogItWithException(WhichLogType.FATAL, "Unable to read " + TrackingFile, ex);
+                WriteToConsoleAndLogItWithException(WhichLogType.FATAL, "Unable to read " + trackingFile, ex);
                 return false;
             }
 
             try
-            { m_Tracking = System.Text.Json.JsonSerializer.Deserialize<Dictionary<string, TrackingInfo>>(FileContents, m_JsonOpts); }
+            { m_Tracking = System.Text.Json.JsonSerializer.Deserialize<Dictionary<string, TrackingInfo>>(fileContents, m_JsonOpts); }
             catch (Exception ex)
             {
-                WriteToConsoleAndLogItWithException(WhichLogType.FATAL, "Unable to deserialize " + TrackingFile, ex);
+                WriteToConsoleAndLogItWithException(WhichLogType.FATAL, "Unable to deserialize " + trackingFile, ex);
                 return false;
             }
 
             return true;
         }
 
-        private static object GetAppSetting(string PropertyName, object DefaultValue, Dictionary<string, object> SettingsDictionary)
+        private static object GetAppSetting(string propertyName, object defaultValue, Dictionary<string, object> settingsDictionary)
         {
-            if (!SettingsDictionary.ContainsKey(PropertyName))
-                return DefaultValue;
+            if (!settingsDictionary.TryGetValue(propertyName, out object value))
+                return defaultValue;
 
-            object temp = SettingsDictionary[PropertyName];
+            object temp = value;
             if (temp is not null and System.Text.Json.JsonElement)
             {
                 System.Text.Json.JsonElement e = (System.Text.Json.JsonElement)temp;
@@ -280,17 +283,17 @@ namespace Check_NOAA_Outlook
                 };
             }
 
-            return SettingsDictionary[PropertyName];
+            return value;
         }
 
         public int RunMain()
         {
             log = new Logging("Check_NOAA_Outlook.log");
 
-            int Result = RunMain2();
+            int result = RunMain2();
             log.Dispose();
 
-            return Result;
+            return result;
         }
 
         private int RunMain2()
@@ -301,8 +304,8 @@ namespace Check_NOAA_Outlook
             //WriteToConsoleAndLogIt(WhichLogType.INFO, string.Format("Starting program, locations: {0}", m_Locations.Count));
 
             // step 1, download the 3 days HTML and pull the image URLs from them
-            List<string> ImageURLs = GetAllImageURLs();
-            if (ImageURLs == null)
+            List<string> imageURLs = GetAllImageURLs();
+            if (imageURLs == null)
             {
                 WriteToConsoleAndLogIt(WhichLogType.FATAL, "ImageURLs was null");
                 return 3;
@@ -310,10 +313,10 @@ namespace Check_NOAA_Outlook
 
             if (m_IncludeMeso)
             {
-                MesoParseResult MesoResult = ParseMainMesoDiscussion();
-                if (MesoResult != MesoParseResult.Found_Mesos)
+                MesoParseResult mesoResult = ParseMainMesoDiscussion();
+                if (mesoResult != MesoParseResult.Found_Mesos)
                 {
-                    if (MesoResult == MesoParseResult.Had_Error) WriteToConsoleAndLogIt(WhichLogType.ERROR, "Had Error parsing main meso discussion");
+                    if (mesoResult == MesoParseResult.Had_Error) WriteToConsoleAndLogIt(WhichLogType.ERROR, "Had Error parsing main meso discussion");
                     else WriteToConsoleAndLogIt(WhichLogType.INFO, "No meso discussions present");
                     m_IncludeMeso = false;
                 }
@@ -322,144 +325,144 @@ namespace Check_NOAA_Outlook
             }
 
             // step 2, download the GIFs for each of the outlook days
-            List<MagickImage> AllTheGifs = DownloadAllTheGIFs(ImageURLs);
-            if (AllTheGifs == null)
+            List<MagickImage> allTheGifs = DownloadAllTheGIFs(imageURLs);
+            if (allTheGifs == null)
             {
                 WriteToConsoleAndLogIt(WhichLogType.FATAL, "AllTheGIFs was null");
                 return 4;
             }
 
-            m_Tracking2 = new Dictionary<string, TrackingInfo>();
+            m_Tracking2 = [];
             // step 3, loop thru each defined location and compare the coordinates from each to the 
             // colors on the downloaded GIFs
             for (int i = 0; i < m_Locations.Count; i++)
             {
-                LocationsInfo LI = m_Locations[i];
-                WriteToConsoleAndLogIt(WhichLogType.DEBUG, "Checking " + LI.Label);
+                LocationsInfo li = m_Locations[i];
+                WriteToConsoleAndLogIt(WhichLogType.DEBUG, "Checking " + li.Label);
 
                 // ------------------------------ here is our payload logic: ------------------------------
-                OutlookResults Results = ParseTheGIFs(AllTheGifs, LI);
-                bool UseHiPri = false;
-                if ((int)Results.HowSevere >= m_Email_HiPri_MinSev)
-                    UseHiPri = true;
+                OutlookResults results = ParseTheGIFs(allTheGifs, li);
+                bool useHiPri = false;
+                if ((int)results.howSevere >= m_Email_HiPri_MinSev)
+                    useHiPri = true;
 
                 // pull previous found info from tracking.xml for comparison
-                PreviousResultsInfo PRI = GetPreviousResults(LI);
+                PreviousResultsInfo pri = GetPreviousResults(li);
 
                 // compare previous results to see if we've already notified for these severities
                 // or if the sevs have changed, email on that
                 //bool HasPrevResults = CheckPreviousResults(PRI, Results, out SeverityChangedFromPrev, out HighestSevFromPrev, out DayChanged);
-                ComparisonAgainPrevResults PrevComparison = CheckPreviousResults(PRI, Results, LI);
+                ComparisonAgainPrevResults prevComparison = CheckPreviousResults(pri, results, li);
                 //bool HasPrevResults = CheckPreviousResults(PRI, Results, LI, out SeverityChangedFromPrev, out HighestSevFromPrev, out DayChanged);
-                bool HasPrevResults = PrevComparison != null;
-                if (!HasPrevResults) PrevComparison = new ComparisonAgainPrevResults();
+                bool hasPrevResults = prevComparison != null;
+                if (!hasPrevResults) prevComparison = new ComparisonAgainPrevResults();
 
-                bool EmailGotSent = false;
+                bool emailGotSent = false;
                 if (!log.IsDebugEnabled)
                 {
-                    StringBuilder sb = new(LI.Label);
-                    sb.AppendFormat(" ({0})", LI.LocationId);
+                    StringBuilder sb = new(li.Label);
+                    sb.AppendFormat(" ({0})", li.locationId);
 
-                    foreach (int Day in Results.Severities.Keys)
+                    foreach (int day in results.severities.Keys)
                     {
-                        SeveritiesEnum Sev = Results.Severities[Day];
-                        sb.AppendFormat(", d{0}: {1}", Day + 1, Sev);
+                        SeveritiesEnum sev = results.severities[day];
+                        sb.AppendFormat(", d{0}: {1}", day + 1, sev);
                     }
                     WriteToConsoleAndLogIt(WhichLogType.INFO, sb.ToString());
                 }
 
-                if (HasPrevResults && PrevComparison.SeverityChangedFromPrev && PrevComparison.HighestSev < LI.Min_Severity)
+                if (hasPrevResults && prevComparison.severityChangedFromPrev && prevComparison.highestSev < li.Min_Severity)
                 {
                     WriteToConsoleAndLogIt(WhichLogType.DEBUG,
-                        "Got sev changes but min sev didn't meet requirements of at least " + LI.Min_Severity.ToString());
+                        "Got sev changes but min sev didn't meet requirements of at least " + li.Min_Severity.ToString());
                 }
 
 #if DEBUG
                 //SeverityChangedFromPrev = true;
 #endif
 
-                List<MesoDiscussionDetails> ApplicableMesos = null;
-                bool MesoChanged = false;
-                if (m_IncludeMeso) ApplicableMesos = GetMesoDiscussionForLocation(LI, PRI, out MesoChanged);
-                bool HasMesos = ApplicableMesos != null && ApplicableMesos.Count > 0;
-                string MesoEmail = null;
+                List<MesoDiscussionDetails> applicableMesos = null;
+                bool mesoChanged = false;
+                if (m_IncludeMeso) applicableMesos = GetMesoDiscussionForLocation(li, pri, out mesoChanged);
+                bool hasMesos = applicableMesos != null && applicableMesos.Count > 0;
+                string mesoEmail = null;
 
-                if (HasMesos)
+                if (hasMesos)
                 {
-                    Results.AttachedImages ??= new Dictionary<Guid, MagickImage>();
-                    for (int t = 0; t < ApplicableMesos.Count; t++)
+                    results.attachedImages ??= [];
+                    for (int t = 0; t < applicableMesos.Count; t++)
                     {
-                        MesoDiscussionDetails Detail = ApplicableMesos[t];
-                        Results.AttachedImages.Add(Detail.Id, Detail.MesoPicture);
-                        if (t == 0) MesoEmail = "\r\n<p><strong>Mesoscale Discussion Below:</strong></p>";
-                        MesoEmail += string.Format("\r\n<p><img src=cid:{0} ></p>\r\n<p>{1}</p>", Detail.Id, Detail.MesoText.Replace(
+                        MesoDiscussionDetails detail = applicableMesos[t];
+                        results.attachedImages.Add(detail.id, detail.mesoPicture);
+                        if (t == 0) mesoEmail = "\r\n<p><strong>Mesoscale Discussion Below:</strong></p>";
+                        mesoEmail += string.Format("\r\n<p><img src=cid:{0} ></p>\r\n<p>{1}</p>", detail.id, detail.mesoText.Replace(
                             "\r\n\r\n", "</p>\r\n\r\n<p>")); //.Replace("\r\n", "<br>\r\n"));
                     }
-                    UseHiPri = true;
+                    useHiPri = true;
                 }
 
                 WriteToConsoleAndLogIt(WhichLogType.DEBUG, string.Format(
                     "HasPrev: {0}, HowSevere: {1}, HighestSevFromPrev: {2}, MinSev: {3}, SevChangedFromPrev: {4}, MesoChanged: {5}, DayChanged: {6}",
-                     HasPrevResults, (int)Results.HowSevere, PrevComparison.HighestSev, LI.Min_Severity,
-                     PrevComparison.SeverityChangedFromPrev, MesoChanged, PrevComparison.DayChanged));
+                     hasPrevResults, (int)results.howSevere, prevComparison.highestSev, li.Min_Severity,
+                     prevComparison.severityChangedFromPrev, mesoChanged, prevComparison.dayChanged));
 
-                if ((int)Results.HowSevere >= LI.Min_Severity)
+                if ((int)results.howSevere >= li.Min_Severity)
                     m_LocationsInSevere++;
 
-                if ((PrevComparison.DayChanged || !HasPrevResults) && (int)Results.HowSevere >= LI.Min_Severity ||
-                    PrevComparison.SeverityChangedFromPrev && PrevComparison.HighestSev >= LI.Min_Severity || MesoChanged)
+                if ((prevComparison.dayChanged || !hasPrevResults) && (int)results.howSevere >= li.Min_Severity ||
+                    prevComparison.severityChangedFromPrev && prevComparison.highestSev >= li.Min_Severity || mesoChanged)
                 {
-                    string EmailSubject = "Severe Weather Forecast in " + LI.Label;
-                    if (PrevComparison.SeverityChangedFromPrev)
+                    string emailSubject = "Severe Weather Forecast in " + li.Label;
+                    if (prevComparison.severityChangedFromPrev)
                     {
-                        WriteToConsoleAndLogIt(WhichLogType.INFO, "Sev changed from last run, sending email, l: {0}", LI.LocationId);
-                        EmailSubject = string.Format("Severe Weather Forecast in {0} Has Changed!", LI.Label);
+                        WriteToConsoleAndLogIt(WhichLogType.INFO, "Sev changed from last run, sending email, l: {0}", li.locationId);
+                        emailSubject = string.Format("Severe Weather Forecast in {0} Has Changed!", li.Label);
                     }
 
-                    if (HasMesos)
+                    if (hasMesos)
                     {
-                        EmailSubject += " Mesoscale Discussion Included";
+                        emailSubject += " Mesoscale Discussion Included";
                     }
 
                     // -------------------------------------- emails get sent here ------------------------------------
                     if (m_EmailIsEnabled)
                     {
-                        EmailGotSent = SendEmail(EmailSubject, Results.EmailMessage, Results.AttachedImages, UseHiPri, LI, MesoEmail);
+                        emailGotSent = SendEmail(emailSubject, results.emailMessage, results.attachedImages, useHiPri, li, mesoEmail);
                         m_LocationsEmailed++;
-                        if (HasMesos) m_MesosIncluded++;
+                        if (hasMesos) m_MesosIncluded++;
                     }
                     else
                         WriteToConsoleAndLogIt(WhichLogType.INFO, "I wanted to send an email but emails are disabled");
                     //EmailGotSent = true;
                 }
-                else if (HasPrevResults && (int)Results.HowSevere > LI.Min_Severity)
+                else if (hasPrevResults && (int)results.howSevere > li.Min_Severity)
                 {
                     WriteToConsoleAndLogIt(WhichLogType.DEBUG, "Not sending email due to previous results");
                 }
 
-                PRI ??= new PreviousResultsInfo(LI.LocationId);
+                pri ??= new PreviousResultsInfo(li.locationId);
 
-                PRI.LastSeverities = Results.Severities;
-                if (HasMesos)
+                pri.lastSeverities = results.severities;
+                if (hasMesos)
                 {
-                    PRI.Mesos = new List<string>();
-                    for (int t = 0; t < ApplicableMesos.Count; t++)
-                        PRI.Mesos.Add(ApplicableMesos[t].Filename);
+                    pri.mesos = [];
+                    for (int t = 0; t < applicableMesos.Count; t++)
+                        pri.mesos.Add(applicableMesos[t].filename);
                 }
                 else
-                    PRI.Mesos = null;
+                    pri.mesos = null;
 
-                AddDataForNextTime(PRI);
+                AddDataForNextTime(pri);
             } // next location
 
             //m_DS_Tracking.WriteXml(TrackingFile);
 
-            string TrackingFile = m_strCD + "Tracking.json";
-            string TrackingText = System.Text.Json.JsonSerializer.Serialize(m_Tracking2, m_JsonOpts);
+            string trackingFile = m_strCD + "Tracking.json";
+            string trackingText = System.Text.Json.JsonSerializer.Serialize(m_Tracking2, m_JsonOpts);
             try
-            { System.IO.File.WriteAllText(TrackingFile, TrackingText); }
+            { System.IO.File.WriteAllText(trackingFile, trackingText); }
             catch (Exception ex)
-            { WriteToConsoleAndLogItWithException(WhichLogType.ERROR, "Unable to write tracking file: " + TrackingFile, ex); }
+            { WriteToConsoleAndLogItWithException(WhichLogType.ERROR, "Unable to write tracking file: " + trackingFile, ex); }
 
             WriteToConsoleAndLogIt(WhichLogType.INFO, "Locations count: {0}, locations in severe: {1}, locations emailed: {2}, mesos included: {3}",
                  m_Locations.Count, m_LocationsInSevere, m_LocationsEmailed, m_MesosIncluded);
@@ -469,7 +472,7 @@ namespace Check_NOAA_Outlook
 
         private MesoParseResult ParseMainMesoDiscussion()
         {
-            m_AllMesoDetails = new Dictionary<Guid, MesoDiscussionDetails>();
+            m_AllMesoDetails = [];
 
             if (m_MesoHtml == null)
             {
@@ -477,38 +480,38 @@ namespace Check_NOAA_Outlook
                 return MesoParseResult.Had_Error;
             }
 
-            string Html = m_MesoHtml;
-            const StringComparison SC = StringComparison.CurrentCultureIgnoreCase;
+            string html = m_MesoHtml;
+            const StringComparison sc = StringComparison.CurrentCultureIgnoreCase;
 
-            int StartingPt = Html.IndexOf("<!-- Contents below-->", SC);
-            if (StartingPt < 1)
+            int startingPt = html.IndexOf("<!-- Contents below-->", sc);
+            if (startingPt < 1)
             {
                 WriteToConsoleAndLogIt(WhichLogType.DEBUG, "Could not find starting point (0)");
                 return MesoParseResult.Had_Error;
             }
 
-            int NoMesos = Html.IndexOf("<center>No Mesoscale Discussions are currently in effect.</center>", StartingPt, SC);
-            if (NoMesos > 0)
+            int noMesos = html.IndexOf("<center>No Mesoscale Discussions are currently in effect.</center>", startingPt, sc);
+            if (noMesos > 0)
             {
                 WriteToConsoleAndLogIt(WhichLogType.INFO, "No Mesoscale Discussions are currently in effect");
                 return MesoParseResult.No_Meso;
             }
 
-            StartingPt = Html.IndexOf("<map name=\"mdimgmap\">", SC);
-            if (StartingPt < 1)
+            startingPt = html.IndexOf("<map name=\"mdimgmap\">", sc);
+            if (startingPt < 1)
             {
                 WriteToConsoleAndLogIt(WhichLogType.DEBUG, "Could not find starting point (2)");
                 return MesoParseResult.Had_Error;
             }
 
-            int EndPoint = Html.IndexOf("</map>", StartingPt + 21, SC);
-            if (EndPoint < 1)
+            int endPoint = html.IndexOf("</map>", startingPt + 21, sc);
+            if (endPoint < 1)
             {
                 WriteToConsoleAndLogIt(WhichLogType.DEBUG, "Could not find end point");
                 return MesoParseResult.Had_Error;
             }
 
-            Html = Html.Substring(StartingPt + 21, EndPoint - StartingPt - 21);
+            html = html.Substring(startingPt + 21, endPoint - startingPt - 21);
             /*
              <map name="mdimgmap">
  <area shape="poly" coords="463,122,475,125,494,128,493,138,483,142,462,141,453,130,463,122,463,122" href="https://www.spc.noaa.gov/products/md/md0302.html" title="Mesoscale Discussion # 302">
@@ -516,70 +519,70 @@ namespace Check_NOAA_Outlook
  </map>
             */
 
-            const string RegexStr = @"(\""poly\"")(.*?)(coords=\"")(.*?)(\"")(.*?)(href=\"")(.*?)(\"")";
-            RegexOptions Opts = RegexOptions.IgnoreCase | RegexOptions.Multiline;
-            MatchCollection Matches = Regex.Matches(Html, RegexStr, Opts);
+            const string regexStr = @"(\""poly\"")(.*?)(coords=\"")(.*?)(\"")(.*?)(href=\"")(.*?)(\"")";
+            RegexOptions opts = RegexOptions.IgnoreCase | RegexOptions.Multiline;
+            MatchCollection matches = Regex.Matches(html, regexStr, opts);
 
-            if (Matches == null || Matches.Count == 0)
+            if (matches == null || matches.Count == 0)
             {
                 WriteToConsoleAndLogIt(WhichLogType.DEBUG, "No regex matches for foly poly coordinates");
                 return MesoParseResult.Had_Error;
             }
 
             //WriteToConsoleAndLogIt(WhichLogType.DEBUG, "Match count: " + Matches.Count.ToString());
-            for (int i = 0; i < Matches.Count; i++)
+            for (int i = 0; i < matches.Count; i++)
             {
-                MesoDiscussionDetails Details = new();
+                MesoDiscussionDetails details = new();
 
-                Match M = Matches[i];
+                Match m = matches[i];
 
-                if (M.Groups.Count < 9)
+                if (m.Groups.Count < 9)
                 {
-                    WriteToConsoleAndLogIt(WhichLogType.DEBUG, $"Match {i}, group count was {M.Groups.Count}, expected >= 9");
+                    WriteToConsoleAndLogIt(WhichLogType.DEBUG, $"Match {i}, group count was {m.Groups.Count}, expected >= 9");
                     return MesoParseResult.Had_Error;
                 }
 
-                Details.PolyCoords = M.Groups[4].Value;
-                Details.Filename = M.Groups[8].Value;
+                details.polyCoords = m.Groups[4].Value;
+                details.filename = m.Groups[8].Value;
 
-                if (!string.IsNullOrEmpty(Details.PolyCoords))
+                if (!string.IsNullOrEmpty(details.polyCoords))
                 {
-                    string[] Split = Details.PolyCoords.Split(',');
-                    List<System.Drawing.Point> Points = new(Split.Length);
-                    for (int t = 0; t < Split.Length; t += 2)
+                    string[] split = details.polyCoords.Split(',');
+                    List<System.Drawing.Point> points = new(split.Length);
+                    for (int t = 0; t < split.Length; t += 2)
                     {
-                        Points.Add(new System.Drawing.Point(
-                            Convert.ToInt32(Split[t].Trim()),
-                            Convert.ToInt32(Split[t + 1].Trim())
+                        points.Add(new System.Drawing.Point(
+                            Convert.ToInt32(split[t].Trim()),
+                            Convert.ToInt32(split[t + 1].Trim())
                             ));
                     }
-                    Details.Poly = Points;
+                    details.poly = points;
                 }
 
                 //WriteToConsoleAndLogIt(WhichLogType.DEBUG, (string.Format("{0}: Poly: {1}, link: {2}", i, Poly, Link));
                 //DownloadMesoDiscussion(BaseURL, Link);
 
-                if (Details.Poly != null && Details.Poly.Count > 0)
-                    m_AllMesoDetails.Add(Details.Id, Details);
+                if (details.poly != null && details.poly.Count > 0)
+                    m_AllMesoDetails.Add(details.id, details);
                 else
-                    WriteToConsoleAndLogIt(WhichLogType.INFO, string.Format("Skipping meso {0} as there was no coords", Details.Filename));
+                    WriteToConsoleAndLogIt(WhichLogType.INFO, string.Format("Skipping meso {0} as there was no coords", details.filename));
             }
             return MesoParseResult.Found_Mesos;
         }
 
-        private bool DownloadMesoDiscussionDetails(Guid WhichOne)
+        private bool DownloadMesoDiscussionDetails(Guid whichOne)
         {
-            MesoDiscussionDetails Details = m_AllMesoDetails[WhichOne];
-            string URL = m_MesoBaseURL + Details.Filename;
+            MesoDiscussionDetails details = m_AllMesoDetails[whichOne];
+            string url = m_MesoBaseURL + details.filename;
 
-            WriteToConsoleAndLogIt(WhichLogType.DEBUG, "Downloading Meso: " + URL);
+            WriteToConsoleAndLogIt(WhichLogType.DEBUG, "Downloading Meso: " + url);
 
-            int RetryCount = 0;
-            string Html = null;
+            int retryCount = 0;
+            string html = null;
 
-            while (RetryCount <= m_Download_MaxRetries)
+            while (retryCount <= m_Download_MaxRetries)
             {
-                Task<string> response = m_MainClient.GetStringAsync(URL);
+                Task<string> response = m_MainClient.GetStringAsync(url);
 
                 try
                 { response.Wait(m_Download_Timeout * 1000); }
@@ -587,140 +590,139 @@ namespace Check_NOAA_Outlook
                 {
                     Exception ex2 = Helpers.GetSingleExceptionFromAggregateException(ex);
 
-                    RetryCount++;
-                    if (RetryCount <= m_Download_MaxRetries)
-                        WriteToConsoleAndLogIt(WhichLogType.WARN, "Error downloading {0}, {1}, retry {2}", URL, ex2.Message, RetryCount);
+                    retryCount++;
+                    if (retryCount <= m_Download_MaxRetries)
+                        WriteToConsoleAndLogIt(WhichLogType.WARN, "Error downloading {0}, {1}, retry {2}", url, ex2.Message, retryCount);
                     else
-                        WriteToConsoleAndLogIt(WhichLogType.WARN, "Error downloading {0}, {1}, cancelling", URL, ex2.Message);
+                        WriteToConsoleAndLogIt(WhichLogType.WARN, "Error downloading {0}, {1}, cancelling", url, ex2.Message);
                     continue;
                 }
 
                 if (response.IsCompletedSuccessfully)
                 {
-                    Html = response.Result;
+                    html = response.Result;
                     break;
                 }
                 else
                 {
-                    if (RetryCount <= m_Download_MaxRetries)
+                    if (retryCount <= m_Download_MaxRetries)
                     {
                         if (response.IsFaulted)
                             WriteToConsoleAndLogIt(WhichLogType.WARN, "Got exception downloading {0}, {1}, retry {1}",
-                                URL, Helpers.GetSingleExceptionFromAggregateException(response.Exception).Message, RetryCount);
+                                url, Helpers.GetSingleExceptionFromAggregateException(response.Exception).Message, retryCount);
                         else
-                            WriteToConsoleAndLogIt(WhichLogType.WARN, "Timed out, retry {0}", RetryCount);
+                            WriteToConsoleAndLogIt(WhichLogType.WARN, "Timed out, retry {0}", retryCount);
                     }
                     else
                     {
                         if (response.IsFaulted)
                             WriteToConsoleAndLogIt(WhichLogType.WARN, "Got exception downloading {0}, {1}, cancelling",
-                                URL, Helpers.GetSingleExceptionFromAggregateException(response.Exception).Message);
+                                url, Helpers.GetSingleExceptionFromAggregateException(response.Exception).Message);
                         else
                             WriteToConsoleAndLogIt(WhichLogType.FATAL, "Timed out, cancelling");
                     }
                 }
             } // next retry
 
-            if (RetryCount > m_Download_MaxRetries)
+            if (retryCount > m_Download_MaxRetries)
                 return false;
 
             // we got the HTML, need to parse it for the the text and image.  Will download image
-            if (!ParseMesoDetails(Details, Html))
+            if (!ParseMesoDetails(details, html))
                 return false;
 
-            List<string> DownloadUrl = new()
-            { m_MesoBaseURL + Details.GifName };
+            List<string> downloadUrl = [m_MesoBaseURL + details.gifName];
 
-            List<MagickImage> Result = DownloadAllTheGIFs(DownloadUrl);
-            if (Result == null || Result.Count == 0)
+            List<MagickImage> result = DownloadAllTheGIFs(downloadUrl);
+            if (result == null || result.Count == 0)
                 return false;
 
-            Details.MesoPicture = Result[0];
-            Details.IsDownloaded = true;
+            details.mesoPicture = result[0];
+            details.isDownloaded = true;
 
             return true;
         }
 
-        private bool ParseMesoDetails(MesoDiscussionDetails Details, string MesoHtml)
+        private bool ParseMesoDetails(MesoDiscussionDetails details, string mesoHtml)
         {
-            string Html = MesoHtml;
-            string RegexStr = @"(<title>)(.*?)(</title>)";
-            RegexOptions Opts = RegexOptions.IgnoreCase | RegexOptions.Multiline;
-            Match Match = Regex.Match(Html, RegexStr, Opts);
+            string html = mesoHtml;
+            string regexStr = @"(<title>)(.*?)(</title>)";
+            RegexOptions opts = RegexOptions.IgnoreCase | RegexOptions.Multiline;
+            Match match = Regex.Match(html, regexStr, opts);
 
-            Details.Title = Details.Filename; // if we can't find the title we'll use the filename
-            if (Match.Success && Match.Groups.Count > 2)
-                Details.Title = Match.Groups[2].Value.Trim();
+            details.title = details.filename; // if we can't find the title we'll use the filename
+            if (match.Success && match.Groups.Count > 2)
+                details.title = match.Groups[2].Value.Trim();
 
-            const StringComparison SC = StringComparison.CurrentCultureIgnoreCase;
-            int StartingPt = Html.IndexOf("a name=\"contents\"", SC);
+            const StringComparison sc = StringComparison.CurrentCultureIgnoreCase;
+            int startingPt = html.IndexOf("a name=\"contents\"", sc);
 
-            if (StartingPt < 1)
+            if (startingPt < 1)
             {
-                WriteToConsoleAndLogIt(WhichLogType.DEBUG, Details.Filename + ", could not find starting point (0)");
+                WriteToConsoleAndLogIt(WhichLogType.DEBUG, details.filename + ", could not find starting point (0)");
                 return false;
             }
 
             // <img src="mcd0330.gif"
             // download gif:
-            RegexStr = @"(img src=\"")(.*?)(\"")";
+            regexStr = @"(img src=\"")(.*?)(\"")";
 
-            Html = Html[(StartingPt + 18)..];
+            html = html[(startingPt + 18)..];
 
-            Match = Regex.Match(Html, RegexStr, Opts);
-            if (!Match.Success)
+            match = Regex.Match(html, regexStr, opts);
+            if (!match.Success)
             {
-                WriteToConsoleAndLogIt(WhichLogType.DEBUG, "Unable to locate gif in html " + Details.Filename);
+                WriteToConsoleAndLogIt(WhichLogType.DEBUG, "Unable to locate gif in html " + details.filename);
                 return false;
             }
 
-            Details.GifName = Match.Groups[2].Value;
+            details.gifName = match.Groups[2].Value;
 
-            StartingPt = Html.IndexOf("<pre>", SC);
+            startingPt = html.IndexOf("<pre>", sc);
 
-            if (StartingPt < 1)
+            if (startingPt < 1)
             {
-                WriteToConsoleAndLogIt(WhichLogType.DEBUG, Details.Filename + ", could not find starting point (1)");
+                WriteToConsoleAndLogIt(WhichLogType.DEBUG, details.filename + ", could not find starting point (1)");
                 return false;
             }
 
-            int EndPt = Html.IndexOf("</pre>", 0, SC);
+            int endPt = html.IndexOf("</pre>", 0, sc);
 
-            if (EndPt < 1)
+            if (endPt < 1)
             {
-                WriteToConsoleAndLogIt(WhichLogType.DEBUG, Details.Filename + ", could not find end point");
+                WriteToConsoleAndLogIt(WhichLogType.DEBUG, details.filename + ", could not find end point");
                 return false;
             }
 
-            Html = Html.Substring(StartingPt + 5, EndPt - StartingPt - 5);
+            html = html.Substring(startingPt + 5, endPt - startingPt - 5);
             //Html = Html.Substring(0, EndPt);
             StringBuilder sb = new();
-            using (System.IO.StringReader SR = new(Html))
+            using (System.IO.StringReader sr = new(html))
             {
-                bool FoundStart = false;
-                while (SR.Peek() >= 0)
+                bool foundStart = false;
+                while (sr.Peek() >= 0)
                 {
-                    string Line = SR.ReadLine();
-                    if (Line == null) continue;
-                    Line = Line.Trim();
+                    string line = sr.ReadLine();
+                    if (line == null) continue;
+                    line = line.Trim();
 
-                    if (!FoundStart)
+                    if (!foundStart)
                     {
-                        if (Line.Length == 0) continue;
-                        FoundStart = true;
+                        if (line.Length == 0) continue;
+                        foundStart = true;
                     }
 
-                    sb.AppendLine(Line);
+                    sb.AppendLine(line);
                 }
             }
 
-            Details.MesoText = sb.ToString();
+            details.mesoText = sb.ToString();
 
             // remove trailing carriage returns, up to 5
             for (int i = 0; i < 5; i++)
             {
-                if (Details.MesoText.EndsWith("\n"))
-                    Details.MesoText = Details.MesoText[0..^1];
+                if (details.mesoText.EndsWith('\n'))
+                    details.mesoText = details.mesoText[0..^1];
                 else
                     break;
             }
@@ -728,63 +730,63 @@ namespace Check_NOAA_Outlook
             return true;
         }
 
-        private List<MesoDiscussionDetails> GetMesoDiscussionForLocation(LocationsInfo LI, PreviousResultsInfo PRI, out bool MesoChanged)
+        private List<MesoDiscussionDetails> GetMesoDiscussionForLocation(LocationsInfo li, PreviousResultsInfo pri, out bool mesoChanged)
         {
-            MesoChanged = false; // only true if there are mesos present now that wasn't before
+            mesoChanged = false; // only true if there are mesos present now that wasn't before
 
             if (m_AllMesoDetails.Count == 0) return null;
-            if (LI.MesoMapLocation.IsEmpty) return null;
+            if (li.mesoMapLocation.IsEmpty) return null;
 
-            List<MesoDiscussionDetails> Mesos = null;
+            List<MesoDiscussionDetails> mesos = null;
 
-            foreach (Guid Id in m_AllMesoDetails.Keys)
+            foreach (Guid id in m_AllMesoDetails.Keys)
             {
-                MesoDiscussionDetails Detail = m_AllMesoDetails[Id];
-                if (Detail.IsPointInPolygon(LI.MesoMapLocation))
+                MesoDiscussionDetails detail = m_AllMesoDetails[id];
+                if (detail.IsPointInPolygon(li.mesoMapLocation))
                 {
                     // found a meso that is for this location
 
                     // first see if we've previously seen this filename
-                    if (PRI == null || PRI.Mesos == null || !PRI.Mesos.Contains(Detail.Filename))
-                        MesoChanged = true;
+                    if (pri == null || pri.mesos == null || !pri.mesos.Contains(detail.filename))
+                        mesoChanged = true;
 
                     WriteToConsoleAndLogIt(WhichLogType.DEBUG, string.Format(
                         "Found applicabled meso for Id: {0}, filename: {1}, IsDownloaded: {2}",
-                        LI.LocationId, Detail.Filename, Detail.IsDownloaded));
-                    if (!Detail.IsDownloaded)
+                        li.locationId, detail.filename, detail.isDownloaded));
+                    if (!detail.isDownloaded)
                     {
-                        bool DownloadedIt = DownloadMesoDiscussionDetails(Id);
-                        if (!DownloadedIt)
+                        bool downloadedIt = DownloadMesoDiscussionDetails(id);
+                        if (!downloadedIt)
                         {
                             WriteToConsoleAndLogIt(WhichLogType.INFO, "Was unable to fully download meso discussion");
                             return null;
                         }
                     }
 
-                    Mesos ??= new List<MesoDiscussionDetails>();
-                    Mesos.Add(Detail);
+                    mesos ??= [];
+                    mesos.Add(detail);
                 }
             }
 
-            return Mesos;
+            return mesos;
         }
 
-        private ComparisonAgainPrevResults CheckPreviousResults(PreviousResultsInfo PRI, OutlookResults OR, LocationsInfo LI)
+        private ComparisonAgainPrevResults CheckPreviousResults(PreviousResultsInfo pri, OutlookResults or, LocationsInfo li)
         {
-            ComparisonAgainPrevResults PrvResults = new();
+            ComparisonAgainPrevResults prvResults = new();
 
-            SeveritiesEnum HighestSevEnum = SeveritiesEnum.Nothing;
-            DateTime Now = DateTime.Now;
-            DateTime UseDate = new(Now.Year, Now.Month, Now.Day, 0, 0, 0, DateTimeKind.Local);
-            DateTime Yesterday = UseDate.AddDays(-1);
-            string[] IOrD = new string[] { "{IncreaseOrDecrease0}", "{IncreaseOrDecrease1}", "{IncreaseOrDecrease2}" };
+            SeveritiesEnum highestSevEnum = SeveritiesEnum.Nothing;
+            DateTime now = DateTime.Now;
+            DateTime useDate = new(now.Year, now.Month, now.Day, 0, 0, 0, DateTimeKind.Local);
+            DateTime yesterday = useDate.AddDays(-1);
+            string[] iOrD = ["{IncreaseOrDecrease0}", "{IncreaseOrDecrease1}", "{IncreaseOrDecrease2}"];
 
-            if (PRI == null || PRI.LastSeverities == null || !PRI.LastCheckTime.HasValue)
+            if (pri == null || pri.lastSeverities == null || !pri.lastCheckTime.HasValue)
             {
-                if (OR.EmailMessage != null)
+                if (or.emailMessage != null)
                 {
-                    foreach (string Text in IOrD)
-                        OR.EmailMessage = OR.EmailMessage.Replace(Text, string.Empty);
+                    foreach (string text in iOrD)
+                        or.emailMessage = or.emailMessage.Replace(text, string.Empty);
                 }
 
                 return null;
@@ -796,229 +798,220 @@ namespace Check_NOAA_Outlook
 
             //bool LastCheckWasPrevDay = false;
 
-            if (PRI.LastCheckTime.HasValue &&
-                PRI.LastCheckTime.Value.ToLocalTime() < UseDate &&
-                PRI.LastCheckTime.Value.ToLocalTime() > Yesterday)
+            if (pri.lastCheckTime.HasValue &&
+                pri.lastCheckTime.Value.ToLocalTime() < useDate &&
+                pri.lastCheckTime.Value.ToLocalTime() > yesterday)
             {
-                PrvResults.DayChanged = true;
+                prvResults.dayChanged = true;
 
                 // we last checked some time yesterday.  Compare the prev day 2 and 3 to see if the severity changed overnight
                 // to do so, we'll shift days prev days 2 and 3 to look like today's 1 and 2
             }
 
-            foreach (int UseDay in PRI.LastSeverities.Keys)
+            foreach (int useDay in pri.lastSeverities.Keys)
             {
-                int UsePrevDay = UseDay;
-                if (PrvResults.DayChanged)
+                int usePrevDay = useDay;
+                if (prvResults.dayChanged)
                 {
-                    if (PRI.LastSeverities.ContainsKey(UsePrevDay + 1))
-                        UsePrevDay++;
+                    if (pri.lastSeverities.ContainsKey(usePrevDay + 1))
+                        usePrevDay++;
                     else
                     {
-                        if (OR.EmailMessage != null)
-                            OR.EmailMessage = OR.EmailMessage.Replace(IOrD[UseDay], string.Empty);
+                        if (or.emailMessage != null)
+                            or.emailMessage = or.emailMessage.Replace(iOrD[useDay], string.Empty);
                         continue;
                     }
                 }
 
-                SeveritiesEnum LastSev = PRI.LastSeverities[UsePrevDay];
-                SeveritiesEnum CurrentSev = OR.Severities[UseDay];
+                SeveritiesEnum lastSev = pri.lastSeverities[usePrevDay];
+                SeveritiesEnum currentSev = or.severities[useDay];
 
-                if (LastSev > HighestSevEnum)
-                    HighestSevEnum = LastSev;
-                if (CurrentSev > HighestSevEnum)
-                    HighestSevEnum = CurrentSev;
+                if (lastSev > highestSevEnum)
+                    highestSevEnum = lastSev;
+                if (currentSev > highestSevEnum)
+                    highestSevEnum = currentSev;
 
-                if (LastSev > CurrentSev &&
-                    (int)LastSev >= LI.Min_Severity)
+                if (lastSev > currentSev &&
+                    (int)lastSev >= li.Min_Severity)
                 {
                     // severity decreased and it was previously at or above the min severity level
-                    PrvResults.SeverityChangedFromPrev = true;
-                    WriteToConsoleAndLogIt(WhichLogType.INFO, "Day {0} Sev decreased from {1} to {2}", UseDay, LastSev, CurrentSev);
-                    OR.EmailMessage = OR.EmailMessage.Replace(IOrD[UseDay],
-                        " (<span style=\"background-color: #FFFF00\">decreased</span> from " + LastSev.ToString() + ")");
+                    prvResults.severityChangedFromPrev = true;
+                    WriteToConsoleAndLogIt(WhichLogType.INFO, "Day {0} Sev decreased from {1} to {2}", useDay, lastSev, currentSev);
+                    or.emailMessage = or.emailMessage.Replace(iOrD[useDay],
+                        " (<span style=\"background-color: #FFFF00\">decreased</span> from " + lastSev.ToString() + ")");
                 }
-                else if (LastSev < CurrentSev &&
-                    (int)CurrentSev >= LI.Min_Severity)
+                else if (lastSev < currentSev &&
+                    (int)currentSev >= li.Min_Severity)
                 {
                     // severity increased and is at or above the min severity level
-                    if (LastSev <= SeveritiesEnum.Nothing) // we don't care if it increased from nothing
+                    if (lastSev <= SeveritiesEnum.Nothing) // we don't care if it increased from nothing
                     {
                         WriteToConsoleAndLogIt(WhichLogType.DEBUG, "Found increase from nothing, so treating it as new");
-                        OR.EmailMessage = OR.EmailMessage.Replace(IOrD[UseDay], string.Empty);
+                        or.emailMessage = or.emailMessage.Replace(iOrD[useDay], string.Empty);
                     }
                     else
                     {
-                        WriteToConsoleAndLogIt(WhichLogType.INFO, "Day {0} Sev increased from {1} to {2}", UseDay, LastSev, CurrentSev);
-                        OR.EmailMessage = OR.EmailMessage.Replace(IOrD[UseDay],
-                            " <span style=\"background-color: #FFFF00\">(increased from " + LastSev.ToString() + ")</span>");
+                        WriteToConsoleAndLogIt(WhichLogType.INFO, "Day {0} Sev increased from {1} to {2}", useDay, lastSev, currentSev);
+                        or.emailMessage = or.emailMessage.Replace(iOrD[useDay],
+                            " <span style=\"background-color: #FFFF00\">(increased from " + lastSev.ToString() + ")</span>");
                     }
-                    PrvResults.SeverityChangedFromPrev = true;
+                    prvResults.severityChangedFromPrev = true;
                 }
                 else
                 {
                     // no change in sev, or it was below thresholds
-                    if (OR.EmailMessage != null)
-                        OR.EmailMessage = OR.EmailMessage.Replace(IOrD[UseDay], string.Empty);
+                    if (or.emailMessage != null)
+                        or.emailMessage = or.emailMessage.Replace(iOrD[useDay], string.Empty);
                 }
             } // next day
 
-            PrvResults.HighestSev = (int)HighestSevEnum;
-            return PrvResults;
+            prvResults.highestSev = (int)highestSevEnum;
+            return prvResults;
         }
 
-        private void AddDataForNextTime(PreviousResultsInfo PRI)
+        private void AddDataForNextTime(PreviousResultsInfo pri)
         {
             //DataView DV = new DataView(m_DS_Tracking.Tables[0]);
             //DV.RowFilter = string.Format("Location_Id = {0}", PRI.LocationId);
 
-            string Mesos = null;
-            if (PRI.Mesos != null && PRI.Mesos.Count > 0)
+            string mesos = null;
+            if (pri.mesos != null && pri.mesos.Count > 0)
             {
-                for (int i = 0; i < PRI.Mesos.Count; i++)
+                for (int i = 0; i < pri.mesos.Count; i++)
                 {
-                    if (i == 0) Mesos = PRI.Mesos[i];
-                    else Mesos += "; " + PRI.Mesos[i];
+                    if (i == 0) mesos = pri.mesos[i];
+                    else mesos += "; " + pri.mesos[i];
                 }
             }
 
             StringBuilder sb = new();
-            foreach (int Day in PRI.LastSeverities.Keys)
+            foreach (int day in pri.lastSeverities.Keys)
             {
                 if (sb.Length > 0) sb.Append(';');
-                sb.Append((int)PRI.LastSeverities[Day]);
+                sb.Append((int)pri.lastSeverities[day]);
             }
 
-            TrackingInfo TI = new()
+            TrackingInfo ti = new()
             {
                 Last_Check = DateTime.UtcNow,
-                Last_Mesos = Mesos,
+                Last_Mesos = mesos,
                 Last_Severities = sb.ToString()
             };
 
-            m_Tracking2.Add(PRI.LocationId, TI);
+            m_Tracking2.Add(pri.locationId, ti);
         }
 
-        private PreviousResultsInfo GetPreviousResults(LocationsInfo LI)
+        private PreviousResultsInfo GetPreviousResults(LocationsInfo li)
         {
             //if (m_DS_Tracking.Tables[0].Rows.Count < 1) return null;
-            if (m_Tracking == null || !m_Tracking.ContainsKey(LI.LocationId)) return null;
+            if (m_Tracking == null || !m_Tracking.TryGetValue(li.locationId, out TrackingInfo value)) return null;
 
-            TrackingInfo TI = m_Tracking[LI.LocationId];
+            TrackingInfo ti = value;
 
-            PreviousResultsInfo PRI = new()
+            PreviousResultsInfo pri = new()
             {
-                LocationId = LI.LocationId
+                locationId = li.locationId
             };
-            if (TI.Last_Check > DateTime.MinValue) PRI.LastCheckTime = TI.Last_Check;
-            if (!string.IsNullOrEmpty(TI.Last_Severities))
+            if (ti.Last_Check > DateTime.MinValue) pri.lastCheckTime = ti.Last_Check;
+            if (!string.IsNullOrEmpty(ti.Last_Severities))
             {
-                string[] Days = TI.Last_Severities.Split(';');
-                PRI.LastSeverities = new Dictionary<int, SeveritiesEnum>();
-                for (int i = 0; i < Days.Length; i++)
+                string[] days = ti.Last_Severities.Split(';');
+                pri.lastSeverities = [];
+                for (int i = 0; i < days.Length; i++)
                 {
-                    string TheDay = Days[i].Trim();
-                    SeveritiesEnum Sev = (SeveritiesEnum)Enum.Parse(typeof(SeveritiesEnum), TheDay);
-                    PRI.LastSeverities.Add(i, Sev);
+                    string theDay = days[i].Trim();
+                    SeveritiesEnum sev = Enum.Parse<SeveritiesEnum>(theDay);
+                    pri.lastSeverities.Add(i, sev);
                 }
             }
-            if (!string.IsNullOrEmpty(TI.Last_Mesos))
+            if (!string.IsNullOrEmpty(ti.Last_Mesos))
             {
-                PRI.Mesos = new List<string>();
-                string[] Temp2 = TI.Last_Mesos.Split(';');
-                for (int i = 0; i < Temp2.Length; i++)
-                    PRI.Mesos.Add(Temp2[i].Trim());
+                pri.mesos = [];
+                string[] temp2 = ti.Last_Mesos.Split(';');
+                for (int i = 0; i < temp2.Length; i++)
+                    pri.mesos.Add(temp2[i].Trim());
             }
 
-            return PRI;
+            return pri;
         }
 
-        private OutlookResults ParseTheGIFs(List<MagickImage> TheGIFs, LocationsInfo LI)
+        private OutlookResults ParseTheGIFs(List<MagickImage> theGIFs, LocationsInfo li)
         {
-            OutlookResults Results = new();
+            OutlookResults results = new();
             StringBuilder sb = new();
-            Results.Severities = new Dictionary<int, SeveritiesEnum>();
+            results.severities = [];
 
-            for (int i = 0; i < TheGIFs.Count; i++)
+            for (int i = 0; i < theGIFs.Count; i++)
             {
-                MagickImage Gif = TheGIFs[i];
+                MagickImage gif = theGIFs[i];
 
-                bool WasSevere = false;
-                SevereCategoryValues SCV = new(Gif, m_Cities);
+                bool wasSevere = false;
+                SevereCategoryValues scv = new(gif, m_Cities);
 
-                System.Drawing.Point UsePoint = new(LI.Location.X, LI.Location.Y);
-                SeveritiesEnum Severity = SCV.WhatSeverityIsThis(UsePoint, false);
-                Results.Severities.Add(i, Severity);
+                System.Drawing.Point usePoint = new(li.location.X, li.location.Y);
+                SeveritiesEnum severity = scv.WhatSeverityIsThis(usePoint, false);
+                results.severities.Add(i, severity);
 
-                if (Severity > Results.HowSevere)
-                    Results.HowSevere = Severity;
+                if (severity > results.howSevere)
+                    results.howSevere = severity;
 
-                int Day = i + 1;
-                Guid UseGuid = Guid.Empty;
+                int day = i + 1;
+                Guid useGuid = Guid.Empty;
 
-                int CheckValue = LI.Min_Severity_For_Picture ?? 2;
-                if ((int)Severity >= CheckValue) // MRGL and up default
+                int checkValue = li.Min_Severity_For_Picture ?? 2;
+                if ((int)severity >= checkValue) // MRGL and up default
                 {
-                    UseGuid = Guid.NewGuid();
-                    MagickImage TheCrop = SCV.GetCropOfMyArea(UsePoint);
-                    Results.AttachedImages ??= new Dictionary<Guid, MagickImage>();
-                    Results.AttachedImages.Add(UseGuid, TheCrop);
-                    Results.WasSevere = true;
-                    WasSevere = true;
+                    useGuid = Guid.NewGuid();
+                    MagickImage theCrop = scv.GetCropOfMyArea(usePoint);
+                    results.attachedImages ??= [];
+                    results.attachedImages.Add(useGuid, theCrop);
+                    results.wasSevere = true;
+                    wasSevere = true;
                 }
 
                 // <img src=""cid:{0}"" />
 
-                string SummaryText = null;
-                if (WasSevere)
+                string summaryText = null;
+                if (wasSevere)
                 {
-                    Results.WasSevere = true;
-                    SummaryText = ExtractSummaryFromSource(i);
+                    results.wasSevere = true;
+                    summaryText = ExtractSummaryFromSource(i);
                 }
 
-                string ImageInfo = WasSevere ?
+                string imageInfo = wasSevere ?
                     string.Format("<br><b>Summary:</b> {2}<br>\r\n<a href=\"{0}\"><img src=cid:{1} ></a>",
-                    m_URLs[i], UseGuid, SummaryText) :
+                    m_URLs[i], useGuid, summaryText) :
                     string.Empty;
 
                 WriteToConsoleAndLogIt(WhichLogType.DEBUG, "Day {0}, Severity: {1}",
-                    Day, Severity);
-                string IncreaseOrDecrease = "{IncreaseOrDecrease" + i.ToString() + "}";
-                DateTime UseDate = DateTime.Now.AddDays(i);
+                    day, severity);
+                string increaseOrDecrease = "{IncreaseOrDecrease" + i.ToString() + "}";
+                DateTime useDate = DateTime.Now.AddDays(i);
                 sb.AppendLine(string.Format("<p><a href=\"{0}\">Day {1}</a> - {2}, Severity: {3}{4}{5}</p>",
-                   m_URLs[i], Day, UseDate.DayOfWeek, Severity, IncreaseOrDecrease, ImageInfo));
+                   m_URLs[i], day, useDate.DayOfWeek, severity, increaseOrDecrease, imageInfo));
             }
 
             //if (Results.WasSevere)
-            Results.EmailMessage = sb.ToString();
+            results.emailMessage = sb.ToString();
 
-            return Results;
+            return results;
         }
 
-        private List<MagickImage> DownloadAllTheGIFs(List<string> URLs)
+        private List<MagickImage> DownloadAllTheGIFs(List<string> urls)
         {
-            List<MagickImage> Results = new();
+            List<MagickImage> results = [];
 
-            List<string> TheURLs = URLs;
-            //if (m_GetExtendedOutlook)
-            //{
-            //    for (int i = 4; i <= 8; i++)
-            //    {
-            //        string Url = m_URL_Extended + string.Format("/day{0}prob.gif", i);
+            List<string> theURLs = urls;
 
-            //        TheURLs.Add(m_URL_Extended);
-            //    }
-            //}
-
-            for (int i = 0; i < TheURLs.Count; i++)
+            for (int i = 0; i < theURLs.Count; i++)
             {
-                string URL = TheURLs[i];
+                string url = theURLs[i];
 
-                WriteToConsoleAndLogIt(WhichLogType.DEBUG, "Downloading: " + URL);
-                int RetryCount = 0;
-                while (RetryCount <= m_Download_MaxRetries)
+                WriteToConsoleAndLogIt(WhichLogType.DEBUG, "Downloading: " + url);
+                int retryCount = 0;
+                while (retryCount <= m_Download_MaxRetries)
                 {
-                    Task<byte[]> response = m_MainClient.GetByteArrayAsync(URL);
+                    Task<byte[]> response = m_MainClient.GetByteArrayAsync(url);
 
                     try
                     { response.Wait(m_Download_Timeout * 1000); }
@@ -1026,11 +1019,11 @@ namespace Check_NOAA_Outlook
                     {
                         Exception ex2 = Helpers.GetSingleExceptionFromAggregateException(ex);
 
-                        RetryCount++;
-                        if (RetryCount <= m_Download_MaxRetries)
-                            WriteToConsoleAndLogIt(WhichLogType.WARN, "Error downloading {0}, {1}, retry {2}", URL, ex2.Message, RetryCount);
+                        retryCount++;
+                        if (retryCount <= m_Download_MaxRetries)
+                            WriteToConsoleAndLogIt(WhichLogType.WARN, "Error downloading {0}, {1}, retry {2}", url, ex2.Message, retryCount);
                         else
-                            WriteToConsoleAndLogIt(WhichLogType.WARN, "Error downloading {0}, {1}, cancelling", URL, ex2.Message);
+                            WriteToConsoleAndLogIt(WhichLogType.WARN, "Error downloading {0}, {1}, cancelling", url, ex2.Message);
                         continue;
                     }
 
@@ -1039,79 +1032,79 @@ namespace Check_NOAA_Outlook
                         try
                         {
                             MagickImage img = new(response.Result);
-                            Results.Add(img);
+                            results.Add(img);
                         }
                         catch (Exception ex)
                         {
-                            WriteToConsoleAndLogItWithException(WhichLogType.ERROR, "Unable to convert to image " + URL, ex);
+                            WriteToConsoleAndLogItWithException(WhichLogType.ERROR, "Unable to convert to image " + url, ex);
                             return null;
                         }
                         break;
                     }
                     else
                     {
-                        RetryCount++;
+                        retryCount++;
 
-                        if (RetryCount <= m_Download_MaxRetries)
+                        if (retryCount <= m_Download_MaxRetries)
                         {
                             if (response.IsFaulted)
                                 WriteToConsoleAndLogIt(WhichLogType.WARN, "Got exception downloading {0}, {1}, retry {1}",
-                                    URL, Helpers.GetSingleExceptionFromAggregateException(response.Exception).Message, RetryCount);
+                                    url, Helpers.GetSingleExceptionFromAggregateException(response.Exception).Message, retryCount);
                             else
-                                WriteToConsoleAndLogIt(WhichLogType.WARN, "Timed out, retry {0}", RetryCount);
+                                WriteToConsoleAndLogIt(WhichLogType.WARN, "Timed out, retry {0}", retryCount);
                         }
                         else
                         {
                             if (response.IsFaulted)
                                 WriteToConsoleAndLogIt(WhichLogType.WARN, "Got exception downloading {0}, {1}, cancelling",
-                                    URL, Helpers.GetSingleExceptionFromAggregateException(response.Exception).Message);
+                                    url, Helpers.GetSingleExceptionFromAggregateException(response.Exception).Message);
                             else
                                 WriteToConsoleAndLogIt(WhichLogType.FATAL, "Timed out, cancelling");
                         }
                     }
                 } // next retry
 
-                if (RetryCount > m_Download_MaxRetries)
+                if (retryCount > m_Download_MaxRetries)
                     return null;
             } // next URL
 
-            return Results;
+            return results;
         }
 
-        private static List<MailboxAddress> ParseEmailRecipients(string Input)
+        private static List<MailboxAddress> ParseEmailRecipients(string input)
         {
-            if (string.IsNullOrEmpty(Input)) return null;
+            if (string.IsNullOrEmpty(input)) return null;
 
-            string[] ToPeople = Input.Split(';');
-            List<MailboxAddress> Addresses = new();
+            string[] toPeople = input.Split(';');
+            List<MailboxAddress> addresses = [];
 
-            for (int i = 0; i < ToPeople.Length; i++)
+            for (int i = 0; i < toPeople.Length; i++)
             {
-                string To = ToPeople[i].Trim();
-                Match M = Regex.Match(To, @"\(.*\)");
-                if (M.Success)
+                string to = toPeople[i].Trim();
+                Match m = RegexTo.Match(to);
+                if (m.Success)
                 {
-                    string Displayname = M.Value[1..^1];
-                    string Email = To[..M.Index].Trim();
+                    string displayname = m.Value[1..^1];
+                    string email = to[..m.Index].Trim();
 
-                    MailboxAddress MA = new(Displayname, Email);
-                    Addresses.Add(MA);
+                    MailboxAddress ma = new(displayname, email);
+                    addresses.Add(ma);
                 }
                 else
-                    Addresses.Add(new MailboxAddress(null, To));
+                    addresses.Add(new MailboxAddress(null, to));
             }
 
-            if (Addresses.Count > 0) return Addresses;
+            if (addresses.Count > 0) return addresses;
             else return null;
         }
 
-        private bool SendEmail(string Subject, string Body, Dictionary<Guid, MagickImage> Images, bool UseHighPri, LocationsInfo LI, string Meso)
+        private bool SendEmail(string subject, string body, Dictionary<Guid, MagickImage> images, bool useHighPri, LocationsInfo li, string meso)
         {
             using SmtpClient client = new();
 
             MimeMessage message = new();
             message.From.Add(new MailboxAddress(null, m_EmailFrom));
-            if (UseHighPri) message.Priority = MessagePriority.Urgent;
+            if (useHighPri) message.Priority = MessagePriority.Urgent;
 
 #if DEBUG
             //if (!string.IsNullOrEmpty(m_EmailTestAddress))
@@ -1120,51 +1113,51 @@ namespace Check_NOAA_Outlook
 
             // optionally email address can have a display name.  Syntax:
             // email@address.com (Display Name)
-            List<MailboxAddress> ToPeople = ParseEmailRecipients(LI.Email_Recipients);
-            if (ToPeople != null)
+            List<MailboxAddress> toPeople = ParseEmailRecipients(li.Email_Recipients);
+            if (toPeople != null)
             {
-                for (int i = 0; i < ToPeople.Count; i++)
-                    message.To.Add(ToPeople[i]);
+                for (int i = 0; i < toPeople.Count; i++)
+                    message.To.Add(toPeople[i]);
             }
 
-            List<MailboxAddress> BCCPeople = ParseEmailRecipients(LI.Email_Recipients_BCC);
-            if (BCCPeople != null)
+            List<MailboxAddress> bccPeople = ParseEmailRecipients(li.Email_Recipients_BCC);
+            if (bccPeople != null)
             {
-                for (int i = 0; i < BCCPeople.Count; i++)
-                    message.Bcc.Add(BCCPeople[i]);
+                for (int i = 0; i < bccPeople.Count; i++)
+                    message.Bcc.Add(bccPeople[i]);
             }
 
-            string Extra = string.Empty;
-            if (Images != null && Images.Count > 0)
+            string extra = string.Empty;
+            if (images != null && images.Count > 0)
             {
                 Guid g = Guid.NewGuid();
 
-                Body += string.Format("\r\n<p><img src=cid:{0} ></p>", g);
-                Images.Add(g, m_Legend);
+                body += string.Format("\r\n<p><img src=cid:{0} ></p>", g);
+                images.Add(g, m_Legend);
             }
 
             if (!string.IsNullOrEmpty(m_EmailReplyTo))
                 message.ReplyTo.Add(new MailboxAddress(null, m_EmailReplyTo));
 
 
-            string Meso2 = Meso ?? string.Empty;
-            string UseBody = "<html><body>\r\n" + Body + "\r\n" + Meso2 + "</body></html>";
+            string meso2 = meso ?? string.Empty;
+            string useBody = "<html><body>\r\n" + body + "\r\n" + meso2 + "</body></html>";
 
-            message.Subject = Subject;
+            message.Subject = subject;
             BodyBuilder bb = new();
 
-            AddImagesToEmail(bb, Images);
+            AddImagesToEmail(bb, images);
 
-            bb.HtmlBody = UseBody;
+            bb.HtmlBody = useBody;
             message.Body = bb.ToMessageBody();
 
 #if DEBUG
-            System.IO.StreamWriter SW = new(@"c:\temp\test.html", false);
-            SW.Write(UseBody);
-            SW.Close();
+            System.IO.StreamWriter sw = new(@"c:\temp\test.html", false);
+            sw.Write(useBody);
+            sw.Close();
 #endif
 
-            bool ItGotSent = false;
+            bool itGotSent = false;
             try
             {
                 client.Connect(m_EmailHost, m_EmailPort, m_UseSSL);
@@ -1173,17 +1166,17 @@ namespace Check_NOAA_Outlook
 
                 client.Send(message);
                 client.Disconnect(true);
-                ItGotSent = true;
+                itGotSent = true;
             }
             catch (Exception ex)
             {
                 WriteToConsoleAndLogItWithException(WhichLogType.ERROR, "Error sending email:", ex);
             }
 
-            if (ItGotSent)
-                WriteToConsoleAndLogIt(WhichLogType.INFO, string.Format("{0} ({1}), Email got sent", LI.Label, LI.LocationId));
+            if (itGotSent)
+                WriteToConsoleAndLogIt(WhichLogType.INFO, string.Format("{0} ({1}), Email got sent", li.Label, li.locationId));
 
-            return ItGotSent;
+            return itGotSent;
         }
 
         private static void AddImagesToEmail(BodyBuilder bb, Dictionary<Guid, MagickImage> images)
@@ -1202,21 +1195,21 @@ namespace Check_NOAA_Outlook
             }
         }
 
-        private string ExtractSummaryFromSource(int WhichDay)
+        private string ExtractSummaryFromSource(int whichDay)
         {
-            string Source = m_SourceCodes[WhichDay];
+            string source = m_SourceCodes[whichDay];
 
-            const StringComparison SC = StringComparison.CurrentCultureIgnoreCase;
-            int Start = Source.IndexOf("...SUMMARY...", SC);
-            if (Start < 0) return null;
+            const StringComparison sc = StringComparison.CurrentCultureIgnoreCase;
+            int start = source.IndexOf("...SUMMARY...", sc);
+            if (start < 0) return null;
 
             StringBuilder sb = new();
-            string Source2 = Source[Start..];
-            System.IO.StringReader SR = new(Source2);
-            SR.ReadLine();
-            while (SR.Peek() > 0)
+            string source2 = source[start..];
+            System.IO.StringReader sr = new(source2);
+            sr.ReadLine();
+            while (sr.Peek() > 0)
             {
-                string Line = SR.ReadLine();
+                string Line = sr.ReadLine();
                 //Line = Line.Substring(27).Trim();
                 // 0        1         2         3
                 // 123456789012345678901234567890
@@ -1232,20 +1225,20 @@ namespace Check_NOAA_Outlook
 
         private List<string> GetAllImageURLs()
         {
-            string URL_Gif = m_NOAA_URL + "day{0}{1}.gif";
+            string url_Gif = m_NOAA_URL + "day{0}{1}.gif";
 
-            List<string> Results = new();
-            List<string> TheURLs = m_URLs;
+            List<string> results = [];
+            List<string> theURLs = m_URLs;
 
-            for (int i = 0; i < TheURLs.Count; i++)
+            for (int i = 0; i < theURLs.Count; i++)
             {
-                string URL = TheURLs[i];
+                string url = theURLs[i];
 
-                WriteToConsoleAndLogIt(WhichLogType.DEBUG, "Downloading: " + URL);
-                int RetryCount = 0;
-                while (RetryCount <= m_Download_MaxRetries)
+                WriteToConsoleAndLogIt(WhichLogType.DEBUG, "Downloading: " + url);
+                int retryCount = 0;
+                while (retryCount <= m_Download_MaxRetries)
                 {
-                    Task<string> response = m_MainClient.GetStringAsync(URL);
+                    Task<string> response = m_MainClient.GetStringAsync(url);
 
                     try
                     { response.Wait(m_Download_Timeout * 1000); }
@@ -1253,171 +1246,171 @@ namespace Check_NOAA_Outlook
                     {
                         Exception ex2 = Helpers.GetSingleExceptionFromAggregateException(ex);
 
-                        RetryCount++;
-                        if (RetryCount <= m_Download_MaxRetries)
-                            WriteToConsoleAndLogIt(WhichLogType.WARN, "Error downloading {0}, {1}, retry {2}", URL, ex2.Message, RetryCount);
+                        retryCount++;
+                        if (retryCount <= m_Download_MaxRetries)
+                            WriteToConsoleAndLogIt(WhichLogType.WARN, "Error downloading {0}, {1}, retry {2}", url, ex2.Message, retryCount);
                         else
-                            WriteToConsoleAndLogIt(WhichLogType.WARN, "Error downloading {0}, {1}, cancelling", URL, ex2.Message);
+                            WriteToConsoleAndLogIt(WhichLogType.WARN, "Error downloading {0}, {1}, cancelling", url, ex2.Message);
                         continue;
                     }
 
                     if (response.IsCompletedSuccessfully)
                     {
-                        string Source = response.Result;
+                        string source = response.Result;
                         if (i < 3)
                         {
                             // day 1-3 outlook
-                            m_SourceCodes[i] = Source;
+                            m_SourceCodes[i] = source;
 
                             // a title="Categorical Outlook"
-                            Match m = Regex.Match(Source, ".*title=.*\"Categorical Outlook\"");
+                            Match m = RegexTitle.Match(source);
                             if (!m.Success) return null;
 
                             // <td OnClick="show_tab('otlk_0100')" OnMouseOver="show_tab('otlk_0100')"><a title="Categorical Outlook"
-                            m = Regex.Match(m.Value, @"\((.*?)\)");
+                            m = RegexOnClick.Match(m.Value);
                             if (!m.Success) return null;
 
                             //('otlk_0100')
-                            string Value = m.Value[2..^2];
+                            string value = m.Value[2..^2];
                             // otlk_0100
 
-                            int UseDay = i + 1;
-                            string UseURL = string.Format(URL_Gif, UseDay, Value);
-                            Results.Add(UseURL);
+                            int useDay = i + 1;
+                            string useURL = string.Format(url_Gif, useDay, value);
+                            results.Add(useURL);
                         }
                         else
                         {
                             // meso discussion
-                            m_MesoHtml = Source;
+                            m_MesoHtml = source;
                         }
 
                         break;
                     }
                     else
                     {
-                        RetryCount++;
+                        retryCount++;
 
-                        if (RetryCount <= m_Download_MaxRetries)
+                        if (retryCount <= m_Download_MaxRetries)
                         {
                             if (response.IsFaulted)
                                 WriteToConsoleAndLogIt(WhichLogType.WARN, "Got exception downloading {0}, {1}, retry {1}",
-                                    URL, Helpers.GetSingleExceptionFromAggregateException(response.Exception).Message, RetryCount);
+                                    url, Helpers.GetSingleExceptionFromAggregateException(response.Exception).Message, retryCount);
                             else
-                                WriteToConsoleAndLogIt(WhichLogType.WARN, "Timed out, retry {0}", RetryCount);
+                                WriteToConsoleAndLogIt(WhichLogType.WARN, "Timed out, retry {0}", retryCount);
                         }
                         else
                         {
                             if (response.IsFaulted)
                                 WriteToConsoleAndLogIt(WhichLogType.WARN, "Got exception downloading {0}, {1}, cancelling",
-                                    URL, Helpers.GetSingleExceptionFromAggregateException(response.Exception).Message);
+                                    url, Helpers.GetSingleExceptionFromAggregateException(response.Exception).Message);
                             else
                                 WriteToConsoleAndLogIt(WhichLogType.FATAL, "Timed out, cancelling");
                         }
                     }
                 } // next retry
 
-                if (RetryCount > m_Download_MaxRetries)
+                if (retryCount > m_Download_MaxRetries)
                     return null;
             }
 
-            return Results;
+            return results;
         }
 
-        private void WriteToConsoleAndLogIt(WhichLogType Logtype, string Text)
+        private void WriteToConsoleAndLogIt(WhichLogType logtype, string text)
         {
-            switch (Logtype)
+            switch (logtype)
             {
                 case WhichLogType.DEBUG:
-                    log.Debug(Text);
-                    if (log.IsDebugEnabled) Console.WriteLine(Text);
+                    log.Debug(text);
+                    if (log.IsDebugEnabled) Console.WriteLine(text);
                     break;
                 case WhichLogType.ERROR:
-                    log.Error(Text);
-                    if (log.IsErrorEnabled) Console.WriteLine(Text);
+                    log.Error(text);
+                    if (log.IsErrorEnabled) Console.WriteLine(text);
                     break;
                 case WhichLogType.FATAL:
-                    log.Fatal(Text);
-                    if (log.IsFatalEnabled) Console.WriteLine(Text);
+                    log.Fatal(text);
+                    if (log.IsFatalEnabled) Console.WriteLine(text);
                     break;
                 case WhichLogType.INFO:
-                    log.Info(Text);
-                    if (log.IsInfoEnabled) Console.WriteLine(Text);
+                    log.Info(text);
+                    if (log.IsInfoEnabled) Console.WriteLine(text);
                     break;
                 case WhichLogType.WARN:
-                    log.Warn(Text);
-                    if (log.IsWarnEnabled) Console.WriteLine(Text);
+                    log.Warn(text);
+                    if (log.IsWarnEnabled) Console.WriteLine(text);
                     break;
             }
         }
 
-        private void WriteToConsoleAndLogIt(WhichLogType Logtype, string Text, params object[] args)
+        private void WriteToConsoleAndLogIt(WhichLogType logtype, string text, params object[] args)
         {
-            switch (Logtype)
+            switch (logtype)
             {
                 case WhichLogType.DEBUG:
-                    log.DebugFormat(Text, args);
-                    if (log.IsDebugEnabled) Console.WriteLine(Text, args);
+                    log.DebugFormat(text, args);
+                    if (log.IsDebugEnabled) Console.WriteLine(text, args);
                     break;
                 case WhichLogType.ERROR:
-                    log.ErrorFormat(Text, args);
-                    if (log.IsErrorEnabled) Console.WriteLine(Text, args);
+                    log.ErrorFormat(text, args);
+                    if (log.IsErrorEnabled) Console.WriteLine(text, args);
                     break;
                 case WhichLogType.FATAL:
-                    log.FatalFormat(Text, args);
-                    if (log.IsFatalEnabled) Console.WriteLine(Text, args);
+                    log.FatalFormat(text, args);
+                    if (log.IsFatalEnabled) Console.WriteLine(text, args);
                     break;
                 case WhichLogType.INFO:
-                    log.InfoFormat(Text, args);
-                    if (log.IsInfoEnabled) Console.WriteLine(Text, args);
+                    log.InfoFormat(text, args);
+                    if (log.IsInfoEnabled) Console.WriteLine(text, args);
                     break;
                 case WhichLogType.WARN:
-                    log.WarnFormat(Text, args);
-                    if (log.IsWarnEnabled) Console.WriteLine(Text, args);
+                    log.WarnFormat(text, args);
+                    if (log.IsWarnEnabled) Console.WriteLine(text, args);
                     break;
             }
         }
 
-        private void WriteToConsoleAndLogItWithException(WhichLogType Logtype, string Text, Exception ex)
+        private void WriteToConsoleAndLogItWithException(WhichLogType logtype, string text, Exception ex)
         {
-            switch (Logtype)
+            switch (logtype)
             {
                 case WhichLogType.DEBUG:
-                    log.Debug(Text, ex);
+                    log.Debug(text, ex);
                     if (log.IsDebugEnabled)
                     {
-                        Console.WriteLine(Text);
+                        Console.WriteLine(text);
                         Console.WriteLine("Exception: " + ex.Message);
                     }
                     break;
                 case WhichLogType.ERROR:
-                    log.Error(Text, ex);
+                    log.Error(text, ex);
                     if (log.IsErrorEnabled)
                     {
-                        Console.WriteLine(Text);
+                        Console.WriteLine(text);
                         Console.WriteLine("Exception: " + ex.Message);
                     }
                     break;
                 case WhichLogType.FATAL:
-                    log.Fatal(Text, ex);
+                    log.Fatal(text, ex);
                     if (log.IsFatalEnabled)
                     {
-                        Console.WriteLine(Text);
+                        Console.WriteLine(text);
                         Console.WriteLine("Exception: " + ex.Message);
                     }
                     break;
                 case WhichLogType.INFO:
-                    log.Info(Text, ex);
+                    log.Info(text, ex);
                     if (log.IsInfoEnabled)
                     {
-                        Console.WriteLine(Text);
+                        Console.WriteLine(text);
                         Console.WriteLine("Exception: " + ex.Message);
                     }
                     break;
                 case WhichLogType.WARN:
-                    log.Warn(Text, ex);
+                    log.Warn(text, ex);
                     if (log.IsWarnEnabled)
                     {
-                        Console.WriteLine(Text);
+                        Console.WriteLine(text);
                         Console.WriteLine("Exception: " + ex.Message);
                     }
                     break;
